Use git_repository instead of http_archive for bazel-bats download
We want to peg to a commit ID anyways, and this way you can simply set the commit
ID without needing an explicit release.

Since we don't really package anything except for source files, there is no
advantage to downloading an archive instead. It is a bit more labor intensive
to clone a git repo, but this is a small repo so it shouldn't be an issue.